ui/templateInfoView: prepare url download sooner and not when creating view
The download may be long, so it makes more sense to prepare it before
creating templateInfoView. For now, there is no functional change but this
allows to prepare the image rendering optimization which will be done in the
next commits.

Note that the TemplateInfoView deleter has been added otherwise we get
on Windows "C2338: can't delete an incomplete type" error.